Reorganized commands into their subsystems
