+++ src/org/usfirst/frc/team1024/robot/commandgroups/AutoSwitchFront.java
@@ -2,10 +2,10 @@ package org.usfirst.frc.team1024.robot.commandgroups;
 import edu.wpi.first.wpilibj.command.CommandGroup;
 import java.lang.Math;
 
-import org.usfirst.frc.team1024.robot.commands.DriveStraight;
-import org.usfirst.frc.team1024.robot.commands.TurnRelative;
 import org.usfirst.frc.team1024.robot.Constants;
 import org.usfirst.frc.team1024.robot.Robot;
+import org.usfirst.frc.team1024.robot.commands.Drive.DriveStraight;
+import org.usfirst.frc.team1024.robot.commands.Drive.TurnRelative;
 /**
  *
  */
