+++ src/org/usfirst/frc/team1024/robot/commandgroups/AutoSwitchMiddle.java
@@ -1,8 +1,8 @@
 package org.usfirst.frc.team1024.robot.commandgroups;
 
 import org.usfirst.frc.team1024.robot.Constants;
-import org.usfirst.frc.team1024.robot.commands.DriveStraight;
-import org.usfirst.frc.team1024.robot.commands.TurnRelative;
+import org.usfirst.frc.team1024.robot.commands.Drive.DriveStraight;
+import org.usfirst.frc.team1024.robot.commands.Drive.TurnRelative;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
