+++ src/org/usfirst/frc/team1024/robot/commandgroups/DriveAndTurn.java
@@ -1,8 +1,8 @@
 package org.usfirst.frc.team1024.robot.commandgroups;
 
 import org.usfirst.frc.team1024.robot.Robot;
-import org.usfirst.frc.team1024.robot.commands.DriveStraight;
-import org.usfirst.frc.team1024.robot.commands.TurnRight;
+import org.usfirst.frc.team1024.robot.commands.Drive.DriveStraight;
+import org.usfirst.frc.team1024.robot.commands.Drive.TurnRight;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
