+++ src/org/usfirst/frc/team1024/robot/commands/Drive/ChangeDriveRamp.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team1024.robot.commands;
+package org.usfirst.frc.team1024.robot.commands.Drive;
 
 import org.usfirst.frc.team1024.robot.Robot;
 
@@ -7,14 +7,14 @@ import edu.wpi.first.wpilibj.command.Command;
 /**
  *
  */
-public class ShiftLow extends Command {
-
-    public ShiftLow() {
-    	requires(Robot.drivetrain);
+public class ChangeDriveRamp extends Command {
+	double ramp;
+    public ChangeDriveRamp(double ramp) {
+    	this.ramp = ramp;
     }
 
     protected void initialize() {
-    	Robot.drivetrain.shiftLow();
+    	Robot.drivetrain.setRamp(ramp);
     }
 
     protected void execute() {
