+++ src/org/usfirst/frc/team1024/robot/commands/Drive/ShiftHigh.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team1024.robot.commands;
+package org.usfirst.frc.team1024.robot.commands.Drive;
 
 import org.usfirst.frc.team1024.robot.Robot;
 
@@ -7,20 +7,21 @@ import edu.wpi.first.wpilibj.command.Command;
 /**
  *
  */
-public class DriveWithJoysticks extends Command {
+public class ShiftHigh extends Command {
 
-    public DriveWithJoysticks() {
+    public ShiftHigh() {
     	requires(Robot.drivetrain);
     }
+    
     protected void initialize() {
+    	Robot.drivetrain.shiftHigh();
     }
 
     protected void execute() {
-    	Robot.drivetrain.drive(Robot.oi.lJoy.getY(), Robot.oi.rJoy.getY());
     }
 
     protected boolean isFinished() {
-        return false;
+        return true;
     }
 
     protected void end() {
