+++ src/org/usfirst/frc/team1024/robot/commands/Drive/ShiftLow.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team1024.robot.commands;
+package org.usfirst.frc.team1024.robot.commands.Drive;
 
 import org.usfirst.frc.team1024.robot.Robot;
 
@@ -7,15 +7,14 @@ import edu.wpi.first.wpilibj.command.Command;
 /**
  *
  */
-public class ChangeDriveSpeed extends Command {
-	double speed;
-    public ChangeDriveSpeed(double speed) {
+public class ShiftLow extends Command {
+
+    public ShiftLow() {
     	requires(Robot.drivetrain);
-    	this.speed = speed;
     }
 
     protected void initialize() {
-    	Robot.drivetrain.posPID.setOutputRange(-speed, speed);
+    	Robot.drivetrain.shiftLow();
     }
 
     protected void execute() {
