+++ src/org/usfirst/frc/team1024/robot/commands/Drive/TurnRight.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team1024.robot.commands;
+package org.usfirst.frc.team1024.robot.commands.Drive;
 
 public class TurnRight extends TurnRelative {
 	public TurnRight(int targetHeading) {
