+++ src/org/usfirst/frc/team1024/robot/commands/DriveAndMoveLift.java
@@ -2,6 +2,8 @@ package org.usfirst.frc.team1024.robot.commands;
 
 import org.usfirst.frc.team1024.robot.Constants;
 import org.usfirst.frc.team1024.robot.Level;
+import org.usfirst.frc.team1024.robot.commands.Drive.DriveStraight;
+import org.usfirst.frc.team1024.robot.commands.lift.MoveLiftPID;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
