+++ src/org/usfirst/frc/team1024/robot/commands/DriveAndShiftAndLift.java
@@ -1,6 +1,7 @@
 package org.usfirst.frc.team1024.robot.commands;
 
 import org.usfirst.frc.team1024.robot.Level;
+import org.usfirst.frc.team1024.robot.commands.lift.MoveLiftPID;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
