+++ src/org/usfirst/frc/team1024/robot/commands/FastCrossToScale.java
@@ -2,6 +2,9 @@ package org.usfirst.frc.team1024.robot.commands;
 
 import org.usfirst.frc.team1024.robot.Level;
 import org.usfirst.frc.team1024.robot.Robot;
+import org.usfirst.frc.team1024.robot.commands.Drive.DriveStraight;
+import org.usfirst.frc.team1024.robot.commands.Drive.TurnLeft;
+import org.usfirst.frc.team1024.robot.commands.Drive.TurnRight;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
