+++ src/org/usfirst/frc/team1024/robot/commands/StraightForwardSwitch.java
@@ -2,6 +2,8 @@ package org.usfirst.frc.team1024.robot.commands;
 
 import org.usfirst.frc.team1024.robot.Constants;
 import org.usfirst.frc.team1024.robot.Level;
+import org.usfirst.frc.team1024.robot.commands.lift.OpenClamp;
+
 import edu.wpi.first.wpilibj.command.CommandGroup;
 
 /**
