+++ src/org/usfirst/frc/team1024/robot/commands/auto/right/DriveToRightScaleEnd.java
@@ -4,21 +4,21 @@ import org.usfirst.frc.team1024.robot.Constants;
 import org.usfirst.frc.team1024.robot.Level;
 import org.usfirst.frc.team1024.robot.Robot;
 import org.usfirst.frc.team1024.robot.commands.AutoDriveConstants;
-import org.usfirst.frc.team1024.robot.commands.ChangeDriveSpeed;
-import org.usfirst.frc.team1024.robot.commands.ChangeLiftSpeed;
-import org.usfirst.frc.team1024.robot.commands.ChangePosPID;
-import org.usfirst.frc.team1024.robot.commands.ChangeRamp;
-import org.usfirst.frc.team1024.robot.commands.ChangeTurnSpeed;
 import org.usfirst.frc.team1024.robot.commands.DriveAndMoveLift;
 import org.usfirst.frc.team1024.robot.commands.DriveAndShift;
 import org.usfirst.frc.team1024.robot.commands.DriveAndShiftAndLift;
-import org.usfirst.frc.team1024.robot.commands.DriveStraight;
 import org.usfirst.frc.team1024.robot.commands.DriveUntilAndMoveLift;
 import org.usfirst.frc.team1024.robot.commands.DriveUntilAndShiftAndLift;
-import org.usfirst.frc.team1024.robot.commands.DriveUntilPID;
-import org.usfirst.frc.team1024.robot.commands.OpenClamp;
-import org.usfirst.frc.team1024.robot.commands.SetCoast;
-import org.usfirst.frc.team1024.robot.commands.TurnLeft;
+import org.usfirst.frc.team1024.robot.commands.Drive.ChangeDriveRamp;
+import org.usfirst.frc.team1024.robot.commands.Drive.ChangeDriveSpeed;
+import org.usfirst.frc.team1024.robot.commands.Drive.ChangePosPID;
+import org.usfirst.frc.team1024.robot.commands.Drive.ChangeTurnSpeed;
+import org.usfirst.frc.team1024.robot.commands.Drive.DriveStraight;
+import org.usfirst.frc.team1024.robot.commands.Drive.DriveUntilPID;
+import org.usfirst.frc.team1024.robot.commands.Drive.SetCoast;
+import org.usfirst.frc.team1024.robot.commands.Drive.TurnLeft;
+import org.usfirst.frc.team1024.robot.commands.lift.ChangeLiftSpeed;
+import org.usfirst.frc.team1024.robot.commands.lift.OpenClamp;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
