+++ src/org/usfirst/frc/team1024/robot/commands/auto/right/DriveToRightSwitch.java
@@ -2,8 +2,8 @@ package org.usfirst.frc.team1024.robot.commands.auto.right;
 
 import org.usfirst.frc.team1024.robot.Robot;
 import org.usfirst.frc.team1024.robot.commands.AutoDriveConstants;
-import org.usfirst.frc.team1024.robot.commands.DriveStraight;
-import org.usfirst.frc.team1024.robot.commands.TurnLeft;
+import org.usfirst.frc.team1024.robot.commands.Drive.DriveStraight;
+import org.usfirst.frc.team1024.robot.commands.Drive.TurnLeft;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
