+++ src/org/usfirst/frc/team1024/robot/commands/intake/IntakeExtendFlat.java
@@ -1,7 +1,4 @@
-package org.usfirst.frc.team1024.robot.commands;
-
-import org.usfirst.frc.team1024.robot.commands.intake.IntakeExtend;
-import org.usfirst.frc.team1024.robot.commands.intake.IntakeFlat;
+package org.usfirst.frc.team1024.robot.commands.intake;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
